% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gages.r
\name{gages}
\alias{gages}
\title{List of gages available for download.}
\usage{
gages()
}
\value{
A simple table of available gages and information about these gages. Note that the 
\code{StartDate} is for any gage data available for a given gage, and may not be available for 
that entire date range for the parameters listed in \code{gageVars}.
}
\description{
Provides a gage list and basic data about the gages that can be downloaded.
}
\details{
This function takes no arguments and is called simply using \code{gages()}. It is used 
internally within the \code{readGages} function, but might provide some minor interpretative value and 
so is provided here as well.

When using \code{readGages} to download gage data from the US Geological Survey Grand Canyon Montoring
and Research Center website (\url{https://www.gcmrc.gov/discharge_qw_sediment/stations/GCDAMP}), 
gage names can be specified using either the gage \code{Name} or \code{Number} 
specified by this function.

The start date for each gage, and its river mile on the Colorado River 
(or the river mile of its confluence if the gage is on a tributary) are provided as well.
Gages are listed as either \code{Mainstem} (on the Colorado River), 
\code{Tributary} (on a tributary adjacent to the mainstem), or 
\code{"Watershed"} (far up a tributary, generally with another gage closer to the mainstem confluence).
}
\examples{
## Look up available gages.
foo1 <- gages()
}
\seealso{
\code{\link{gageVars}}, which provides a list of available gage parameters.
\code{\link{readGage}}, which allows downloading of these gage data from the website.
}
\author{
Jeffrey D. Muehlbauer, \email{jmuehlbauer@usgs.gov}
}
\concept{database, website, gage, gauge}
