% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gageVars.r
\name{gageVars}
\alias{gageVars}
\title{List of gage parameters available for download.}
\usage{
gageVars()
}
\value{
A simple table of available gage parameters, along with their units.
}
\description{
Provides a list of gage parameters, names, and units.
}
\details{
This function takes no arguments and is called simply using \code{gageVars()}. It is used 
internally within the \code{readGages} function, but might provide some minor interpretative value and 
so is provided here as well.

When using \code{readGages} to download gage data from the US Geological Survey Grand Canyon Montoring
and Research Center website (\url{https://www.gcmrc.gov/discharge_qw_sediment/stations/GCDAMP}), 
gage parameters (variables) can be specified using the \code{Parameter} name specified by this function.
}
\examples{
## Look up available gage parameters.
foo1 <- gageVars()
}
\seealso{
\code{\link{gages}}, which provides a list of available gages.
\code{\link{readGage}}, which allows downloading of these gage data from the website.
}
\author{
Jeffrey D. Muehlbauer, \email{jmuehlbauer@usgs.gov}
}
\concept{database, website, gage, gauge}
