% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGage.r
\name{readGage}
\alias{readGage}
\title{Download Grand Canyon gage data}
\usage{
readGage(gage, vars = "all", startDate = "first", endDate = "today",
  cutCols = TRUE, cutRows = TRUE, writeCSV = FALSE)
}
\arguments{
\item{gage}{Vector of gages to download. See Details.}

\item{vars}{Vector of gage parameters to download. See Details. Default is \code{all}.}

\item{startDate}{The beginning date for the data download (includes this day). See Details. Default is 
\code{first}.}

\item{endDate}{The end date for the data download (includes this day). See Details. Default is 
\code{today}.}

\item{cutCols}{Whether to cut any empty (\code{NA}) columns (gage parameters). Default is \code{TRUE}.}

\item{cutRows}{Whether to cut any empty (\code{NA}) rows of data. The website often does this 
internally regardless. Default is \code{TRUE}.}

\item{writeCSV}{Whether to write the downloaded data to a CSV file. See Details. Default is \code{FALSE}.}
}
\value{
If \code{gage} is a vector of multiple gages, returns a list of dataframes, with one list 
element for each gage. These are sorted in the order specified by \code{gage}, with the list elemetn 
named according to the names specified by \code{gage} as well. If \code{gage} is only a single gage, 
the function returns a single dataframe of gage data (not a list). If no data are available for a gage, 
only \code{NA} is returned. This is provided (rather than an error) to enable batch download of 
multiple gages. 

Note on units: The first column is always \code{DateTime}, which is provided in \code{POSIXlt} format 
in Mountain Standard Time (MST). Units on other columns (parameters) can be found using a call to 
\code{gageVars()}.

Note on processing time: For downloads from a single gage for ~1 year or less of data, download times
should be similar to the time it would take to download a .tsv file directly from the website. For much 
larger datasets or multiple gages, download times can increase dramatically. These will be processed in 
parallel and are generally much faster than downloading directly from the website, but may take
several minutes.
}
\description{
Allows data download from gages in and around the Colorado River in Grand Canyon.
}
\details{
This function enables downloading of gage data from the US Geological Survey Grand Canyon Montoring 
and Research Center website (\url{https://www.gcmrc.gov/discharge_qw_sediment/stations/GCDAMP}). 

Gage names and parameters available for download can be found using a call to \code{gages()} and 
\code{gageVars()}, respectively. Gages can be specified using either the gage \code{Name} or 
\code{Number}.

The \code{startDate} and \code{endDate} specified indicate the first and last day of gage data to be 
dowloaded, includive of those dates. The start date for each gage can be found using a call to 
\code{gages()}. Dates should be formatted as \code{YYYY-MM-DD} (\code{'\%Y-\%m-\%d'} in \code{R} syntax);
shortcuts \code{'first'} and \code{'today'} can also be used to download data from the first date of 
record for each gage, to today.

The arguments \code{cutCols} and \code{cutRows} will truncate the data for each gage to only the 
DateTimes that have any data associated with them. In other words, these arguments will cut out any 
columns and rows that contain only \code{NA}, respectively. Note that the website automatically 
truncates blank rows at the beginning and end of the dataset, so in practice much of this \code{NA} 
removal occurs regardless of the value specified for these arguments.

The argument \code{writeCSV} specifies whether the downloaded gage data should be written to a .csv 
file. If \code{writeCSV = TRUE}, gage data are written to the working directory, with the filename
\code{GageData_NAME.csv}, where \code{NAME} is the name of each gage specified in the \code{gage}
argument. Alternately, instead of \code{TRUE}, \code{writeCSV} can accept a filename or filepath. 
In this case, the gage name will be appended to the character string specified by \code{writeCSV}
(see Examples).
}
\examples{
## Look up available gages.
foo1 <- gages()

## Look up available gage parameters.
foo2 <- gageVars()

## Download discharge and stage data from Lees Ferry (AKA 09380000) and 30-Mile, from 1-2 January, 2002.
foo3 <- readGage(gage = c("09380000", "CR30Mile"), vars = c("Discharge", "Stage"), 
	startDate = "2002-01-01", endDate = "2002-01-02")

## Same as foo3, but write data to a "Data" folder in the working directory.
foo4 <- readGage(gage = c("09380000", "CR30Mile"), vars = c("Discharge", "Stage"), 
	startDate = "2002-01-01", endDate = "2002-01-02", writeCSV = "Data/")
}
\seealso{
\code{\link{gages}}, which provides a list of available gages.
\code{\link{gageVars}}, which provides a list of available gage parameters.
}
\author{
Jeffrey D. Muehlbauer, \email{jmuehlbauer@usgs.gov}
}
\concept{database, website, gage, gauge}
